Extract out the CMS components into multiple files.
