+++ examples/minicms/frontend/src/config.js
@@ -1,0 +1,2 @@
+const  API_PREFIX_URL = 'http://localhost:4000/api/pages';
+export { API_PREFIX_URL };
